third_party/blink/web_tests/plugins/mouse-events-fixedpos-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(true)
CONSOLE MESSAGE: Blink Test Plugin: MouseDown at (50,50)
CONSOLE MESSAGE: Blink Test Plugin: MouseUp at (50,50)
CONSOLE MESSAGE: Blink Test Plugin: MouseDown at (60,60)
CONSOLE MESSAGE: Blink Test Plugin: MouseUp at (70,60)
Tests for widget positions being correctly updated after scrolling. rdar://problem/7559069
